refactor(eval): extract pipeline parsing and tests from mod.rs
Extract YAML parsing logic and tests from the 2235-line pipeline/mod.rs
into dedicated sibling modules. No behavior changes; identical test
results (366 unit tests + 47 others pass).

New/modified files:
  pipeline/mod.rs      -- structs, apply logic, merge/apply functions (448 lines, was 2235)
  pipeline/parsing.rs  -- all YAML parsing and condition parsing (817 lines)
  pipeline/tests.rs    -- all unit tests (982 lines)
